Use the Google Chrome User Profile ...
Instead of creating a new user profile everytime a user wants to pull
the code base, or use it in any capacity.
Use the google chrome default app data.

A lot of debugging and trial and error took place to get this working,
along with some testing regarding shell == true, which would be an issue
on a Server, or if it took user input, but for our purpose. especially
on windows, it should be reasonably safe.

	some thoughts on removing shell == true,

	which obviously had unintended side effects...
	caution about shell == true is out there, understand why before going
	this route ... because this is pretty dumb

complete change

windows will now use localappdata/google/chrome/user data to get the
current users chrome profile.
should make users life a lot easier